Fix minor bug
Hi, 
I am a big fan of your plugin and I use it often. 
I realised there was a minor issue in the code when clicking the 'single layer' icon and selecting a group (and not a layer). 
I fixed it and I would like to share with you, hoping this can help to push anupdate in the QGIS Repo. 

Cheers, 

Val